refactor(core/delizia): use `u8` for header margin
[no changelog]

+++ core/embed/rust/src/ui/layout_delizia/flow/confirm_action.rs
@@ -76,7 +76,7 @@ pub struct ConfirmActionOptions {
     pub swipe_up: bool,
     pub page_limit: Option<u16>,
     pub page_counter: bool,
-    pub frame_margin: usize,
+    pub frame_margin: u8,
 }
 
 impl ConfirmActionOptions {
@@ -116,7 +116,7 @@ impl ConfirmActionOptions {
         self
     }
 
-    pub fn with_frame_margin(mut self, frame_margin: usize) -> Self {
+    pub fn with_frame_margin(mut self, frame_margin: u8) -> Self {
         self.frame_margin = frame_margin;
         self
     }
